# Some performance improvements for the initial refinement