--- a/src/Trixi.jl
+++ b/src/Trixi.jl
@@ -37,7 +37,7 @@ using RecipesBase
 using Requires
 @reexport using StaticArrays: SVector
 using StaticArrays: MVector, MArray, SMatrix
-using TimerOutputs: @notimeit, @timeit_debug, TimerOutput, print_timer, reset_timer!
+using TimerOutputs: TimerOutputs, @notimeit, @timeit_debug, TimerOutput, print_timer, reset_timer!
 using UnPack: @unpack, @pack!
 
 # Tullio.jl makes use of LoopVectorization.jl via Requires.jl.