@ranocha 2021-03-26
Digging a bit deeper shows that most of the time is spent in `insert!`, there in `move!`, where `raw_copy!` and `move_connectivity!` take roughly 50% each. These methods don't allocate and look reasonable optimized to me. Hence, algorithmic changes avoiding this excessive moving and copying of data is necessary to get acceptable performance.