@sloede 2021-03-26
@ranocha Thanks a lot for having a closer look into this issue and for already providing some improvements!

> Hence, algorithmic changes avoiding this excessive moving and copying of data is necessary to get acceptable performance.

Yeah, the excessive moving and copying is part of the "better safe than sorry" strategy I applied when implementing this. My first attempt at optimization would be to allow for non-sorted cells during the initial refinement process, deferring the refinement to after the initial refinement has been achieved. I don't know how many code changes this would entail, since I have to admit that I forget if and where I relied on the depth-first sorting...