@ranocha 2021-03-26
> My first attempt at optimization would be to allow for non-sorted cells during the initial refinement process, deferring the sorting to after the initial refinement has been achieved. I don't know how many code changes this would entail, since I have to admit that I forget if and where I relied on the depth-first sorting...

Yeah, this might need some time to implement, but we should definitely do it soon - to avoid the excessive cost of initializing a simple mesh with exponential (?) complexity...